chore: useFiltersContext and FiltersProvider
