+++ packages/frontend/src/components/common/Filters/FiltersProvider/useFiltersContext.tsx
@@ -1,0 +1,12 @@
+import { useContext } from 'react';
+import { Context, FiltersContext } from '.';
+
+export function useFiltersContext(): FiltersContext {
+    const context = useContext(Context);
+    if (context === undefined) {
+        throw new Error(
+            'useFiltersContext must be used within a FiltersProvider',
+        );
+    }
+    return context;
+}
